docs: ファイルの配置構成を変更 #5
