+++ web_basics/03_Simple_WebServer/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "editor.formatOnSave": true,
+    "workbench.editor.enablePreview": false
+}
