+++ web_basics/03_Simple_WebServer/public/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+
+    <link rel="stylesheet" href="style.css">
+    <script src="index.js"></script>
+</head>
+<body>
+    <h1>Welcome</h1>
+
+    <div>
+        <button>
+            Submit
+        </button>
+    </div>
+
+    <pre>
+        <code>
+            const XHR = new XMLHttpRequest();
+        </code>
+    </pre>
+</body>
+</html>
