+++ web_basics/03_Simple_WebServer/public/index.js
@@ -1,0 +1,1 @@
+console.log("loaded");
