+++ web_basics/03_Simple_WebServer/public/style.css
@@ -1,0 +1,3 @@
+button {
+    color: #e0e0e0;
+}
